+++ src/AndTargetingRule.php
@@ -31,4 +31,10 @@ final readonly class AndTargetingRule implements TargetingRule
 
         return true;
     }
+
+    /** @return list<TargetingRule> */
+    public function getRules(): array
+    {
+        return $this->rules;
+    }
 }